requirements/base.txt
# base.txt
Django==4.1.5
djangorestframework==3.14.0
django-environ==0.9.0
geopandas
plotly_express
osmnx